requirements.txt
tensorflow-gpu>=2.3.0
imgaug==0.4
opencv-python~=4.5
sklearn
matplotlib~=3.3
tqdm
pascal_voc_writer
tf2onnx
flask==1.1.4
flask-socketio==4.3.1
flask_cors==3.0.10
gdown
Jinja2==2.10.1
markupsafe==1.1.1
python-socketio==4.6.0
python-engineio==3.13.2